Added method and class javadoc comments, improved edge case issues and syncing with database. blockchain addblock service returns block now not bool, which is more useful. tested, runs find

+++ Java BE/Case Study/src/main/java/privblock/gerald/ryan/dao/BlockDao.java
@@ -18,7 +18,7 @@ public class BlockDao extends DBConnection implements BlockDaoI {
 	}
 
 	@Override
-	public Block getBlock(int id) {
+	public Block getBlock(long id) {
 		this.connect();
 		Block b = em.find(Block.class, id);
 		this.disconnect();
